Update README.md
Paralysis patients, such as Tetraplegics, suffer greatly due to their physical limitations. To overcome this we are developing a smart home control system that uses eye blink. It captures the signals using an eye blink sensor and the signal generated by the sensor will be received and then transmitted to a microcontroller. In this process, it eliminates the noise and remove the artifacts from the data to represent true signals and the signal processed will be actuated based on the time algorithm. Present systems use expensive controller for the communication and bulky hardware to run the system. So, to overcome this limitation this algorithm eliminates false detection and the missed blinks. Here we are using eye blink sensor to count the number of blinks, based on which applications can be controlled. IR sensor is interfaced with the ESP32 which controls the operation of the applications. Based on the number of counts the device will be turned on or it will be turned off. The complete system is powered from a battery. The project is built around the ESP32 it provides all the functionality and control. This device can also be controlled using Google Assistant on the cell phone.